pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "sentinal"
version = "0.0.1"
authors = [
  { name="J-Palmer", email="redrum519git@gmail.com" },
]
description = "A low-resource system administration CLI for Linux desktop workflows."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "google-adk",
    "requests",
]

[project.scripts]
sentinal = "sentinal.cli:main"

[project.optional-dependencies]
dev = [
    "pytest",
    "fastapi",
    "uvicorn[standard]",
]
gpu = ["pynvml"]
llm = ["llama-cpp-python"]
metrics = ["prometheus-client"]
ebpf = ["bcc"]
full = ["sentinal[gpu,llm,metrics,ebpf]"]

[tool.setuptools]
package-dir = {"" = "src"}
